Consolidate common deployment backend code
…from `KobocatDeploymentBackend` and `MockDeploymentBackend` into
`BaseDeploymentBackend`
