+++ kpi/deployment_backends/base_backend.py
@@ -4,6 +4,7 @@ import abc
 import copy
 import datetime
 import json
+import uuid
 from datetime import date
 from typing import Union, Iterator, Optional
 
@@ -12,6 +13,7 @@ from django.db.models.query import QuerySet
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as t
 from django.core.exceptions import PermissionDenied
+from lxml import etree
 from rest_framework import serializers
 from rest_framework.pagination import _positive_int as positive_int
 from shortuuid import ShortUUID
@@ -19,12 +21,15 @@ from shortuuid import ShortUUID
 from kpi.constants import (
     SUBMISSION_FORMAT_TYPE_XML,
     SUBMISSION_FORMAT_TYPE_JSON,
+    PERM_CHANGE_SUBMISSIONS,
     PERM_PARTIAL_SUBMISSIONS,
     PERM_VIEW_SUBMISSIONS,
 )
+from kpi.exceptions import BulkUpdateSubmissionsClientException
 from kpi.models.asset_file import AssetFile
 from kpi.models.paired_data import PairedData
 from kpi.utils.django_orm_helper import UpdateJSONFieldAttributes
+from kpi.utils.xml import edit_submission_xml
 
 
 class BaseDeploymentBackend(abc.ABC):
@@ -70,11 +75,106 @@ class BaseDeploymentBackend(abc.ABC):
     def bulk_assign_mapped_perms(self):
         pass
 
-    @abc.abstractmethod
     def bulk_update_submissions(
         self, data: dict, user: 'auth.User'
     ) -> dict:
-        pass
+        """
+        Allows for bulk updating (bulk editing) of submissions. A
+        `deprecatedID` for each submission is given the previous value of
+        `instanceID` and `instanceID` receives an updated uuid. For each key
+        and value within `request_data`, either a new element is created on the
+        submission's XML tree, or the existing value is replaced by the updated
+        value.
+
+        Args:
+            data (dict): must contain a list of `submission_ids` and at
+                least one other key:value field for updating the submissions
+            user (User)
+
+        Returns:
+            dict: formatted dict to be passed to a Response object
+        """
+        submission_ids = self.validate_access_with_partial_perms(
+            user=user,
+            perm=PERM_CHANGE_SUBMISSIONS,
+            submission_ids=data['submission_ids'],
+            query=data['query'],
+        )
+
+        # If `submission_ids` is not empty, user has partial permissions.
+        # Otherwise, they have full access.
+        if submission_ids:
+            # Reset query, because all the submission ids have been already
+            # retrieve
+            data['query'] = {}
+        else:
+            submission_ids = data['submission_ids']
+
+        submissions = self.get_submissions(
+            user=user,
+            format_type=SUBMISSION_FORMAT_TYPE_XML,
+            submission_ids=submission_ids,
+            query=data['query'],
+        )
+
+        if not self.current_submission_count:
+            raise BulkUpdateSubmissionsClientException(
+                detail=t('No submissions match the given `submission_ids`')
+            )
+
+        # Remove potentially destructive keys from the payload
+        update_data = copy.deepcopy(data['data'])
+        update_data = {
+            k: v
+            for k, v in update_data.items()
+            if not (
+                k in self.PROTECTED_XML_FIELDS
+                or '/' in k
+                and k.split('/')[0] in self.PROTECTED_XML_FIELDS
+            )
+        }
+
+        kc_responses = []
+        for submission in submissions:
+            xml_parsed = etree.fromstring(submission)
+
+            _uuid, uuid_formatted = self.generate_new_instance_id()
+
+            # Updating xml fields for submission. In order to update an existing
+            # submission, the current `instanceID` must be moved to the value
+            # for `deprecatedID`.
+            instance_id = xml_parsed.find('meta/instanceID')
+            # If the submission has been edited before, it will already contain
+            # a deprecatedID element - otherwise create a new element
+            deprecated_id = xml_parsed.find('meta/deprecatedID')
+            deprecated_id_or_new = (
+                deprecated_id
+                if deprecated_id is not None
+                else etree.SubElement(xml_parsed.find('meta'), 'deprecatedID')
+            )
+            deprecated_id_or_new.text = instance_id.text
+            instance_id.text = uuid_formatted
+
+            # If the form has been updated with new fields and earlier
+            # submissions have been selected as part of the bulk update,
+            # a new element has to be created before a value can be set.
+            # However, with this new power, arbitrary fields can be added
+            # to the XML tree through the API.
+            for path, value in update_data.items():
+                edit_submission_xml(xml_parsed, path, value)
+
+            kc_response = self.store_submission(
+                user, etree.tostring(xml_parsed), _uuid
+            )
+            kc_responses.append(
+                {
+                    'uuid': _uuid,
+                    'response': kc_response,
+                }
+            )
+
+        return self.prepare_bulk_update_response(kc_responses)
+
 
     @abc.abstractmethod
     def calculated_submission_count(self, user: 'auth.User', **kwargs):
@@ -120,6 +220,16 @@ class BaseDeploymentBackend(abc.ABC):
     @abc.abstractmethod
     def enketo_id(self):
         pass
+
+    @staticmethod
+    def generate_new_instance_id() -> (str, str):
+        """
+        Returns:
+            - Generated uuid
+            - Formatted uuid for OpenRosa xml
+        """
+        _uuid = str(uuid.uuid4())
+        return _uuid, f'uuid:{_uuid}'
 
     @abc.abstractmethod
     def get_attachment(
@@ -352,6 +462,13 @@ class BaseDeploymentBackend(abc.ABC):
     @property
     def stored_data_key(self):
         return self.__stored_data_key
+
+    @property
+    @abc.abstractmethod
+    def store_submission(
+        self, user, xml_submission, submission_uuid, attachments=None
+    ):
+        pass
 
     @property
     @abc.abstractmethod
